+++ changelog.md
@@ -4,6 +4,7 @@
 - Make date fields return carbon object
 - Add `loadRelationFields` method
 - Add abstract tags field
+- Add entity duplication functionality 
 
 ## Version 1.8.1 (2017-07-19)
 - Fix template creation from command line