+++ changelog.md
@@ -5,6 +5,8 @@
 - Add `loadRelationFields` method
 - Add abstract tags field
 - Add entity duplication functionality 
+- Add orWhereValue and leftJoinValueByKey scopes to Entity
+- Allow search query to be modified in EntityDecorator
 
 ## Version 1.8.1 (2017-07-19)
 - Fix template creation from command line